// src/app/interfaces/plan.ts
export interface Planes{
  id: string;
  codigo: string;
  nombre: string;
  descripcion_solicitud: string;
  cantidad_personas: string;
  fecha_solicitud: string;
  fecha_inicio_viaje: string;
  fecha_fin_viaje: string;
  valor_total:string;
  fecha_creacion: string;
  fecha_modificacion: string;
  usu_creador: string;
  usu_modificador: string;
  estado: string;
  id_clie: string;
  id_usua: string;
}